+++ react/selectors/fields.ts
@@ -7,14 +7,37 @@ import last from 'lodash/last'
 import { POSTAL_CODE, ONE_LEVEL, TWO_LEVELS, THREE_LEVELS } from '../constants'
 import hasOption from './hasOption'
 import cleanStr from './cleanStr'
+import type {
+  PostalCodeFieldRule,
+  Rule,
+  AddressRules,
+  PostalCodeRules,
+} from '../types/rules'
+import type {
+  Fields,
+  AddressWithValidation,
+  ValidatedField,
+  AddressValues,
+} from '../types/address'
 
-export function getField(fieldName, rules) {
-  return find(rules.fields, ({ name }) => name === fieldName)
+export function getField(fieldName: Fields, rules: AddressRules) {
+  if ('fields' in rules) {
+    return rules.fields.find((field) => field.name === fieldName)
+  }
+
+  return rules[fieldName]
 }
 
-export function hasOptions(field, address) {
-  const hasValueOptions =
-    address && address[field.name] && address[field.name].valueOptions
+export function hasOptions(
+  field: Rule,
+  address?: AddressWithValidation
+): field is PostalCodeFieldRule {
+  // geolocation rules does not contain options
+  if ('valueIn' in field) {
+    return false
+  }
+
+  const hasValueOptions = !!address?.[field.name]?.valueOptions
 
   return !!(
     field.options ||
@@ -24,8 +47,14 @@ export function hasOptions(field, address) {
   )
 }
 
-function getFieldValue(field) {
-  return typeof field === 'object' ? field.value : field
+function getFieldValue(field: ValidatedField | AddressValues): AddressValues {
+  if (field == null) {
+    return undefined
+  }
+
+  return typeof field === 'object' && !Array.isArray(field)
+    ? field.value
+    : field
 }
 
 export function normalizeOptions(options) {
@@ -117,9 +146,28 @@ function getSecondLevelOptions(field, address) {
   return []
 }
 
-function getThirdLevelOptions(field, address, rules) {
-  const secondLevelField = getField(field.basedOn, rules)
-  const firstLevelField = getField(secondLevelField.basedOn, rules)
+function getThirdLevelOptions(
+  field: PostalCodeFieldRule,
+  address: AddressWithValidation,
+  rules: PostalCodeRules
+) {
+  const secondLevelField = getField(
+    field.basedOn!,
+    rules
+  ) as PostalCodeFieldRule
+
+  if (!secondLevelField) {
+    return []
+  }
+
+  const firstLevelField = getField(
+    secondLevelField.basedOn!,
+    rules
+  ) as PostalCodeFieldRule
+
+  if (!firstLevelField) {
+    return []
+  }
 
   const secondLevelValue = getFieldValue(address[secondLevelField.name])
   const firstLevelValue = getFieldValue(address[firstLevelField.name])
@@ -144,7 +192,7 @@ function toValueAndLabel(option) {
 }
 
 export function getDependentFields(fieldName, rules) {
-  let dependentFields = []
+  let dependentFields: string[] = []
 
   if (fieldAffectsPostalCode(fieldName, rules)) {
     dependentFields = [...dependentFields, 'postalCode']